pyproject.toml
[tool.poetry]
name = "ontology-rag"
version = "0.1.0"
description = "Simple RAG system for ontology learning and experimentation"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ontology_rag", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
chromadb = "^1.2.0"
httpx = "^0.27.0"
typer = "^0.12.0"
rich = "^13.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
ontology-rag = "ontology_rag.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
